Move Chart.ViewContainer into own file.
